Registrando e excluindo chaves Pix no Banco Central (BCB)

+++ src/main/kotlin/br/com/zup/pix/actions/remove/RemoveChaveService.kt
@@ -1,12 +1,16 @@
 package br.com.zup.pix.actions.remove
 
 import br.com.zup.RemoveChavePixRequest
+import br.com.zup.integration.bcb.BcbClient
+import br.com.zup.integration.bcb.DeletaChaveBcbRequest
 import br.com.zup.integration.itau.ItauClient
 import br.com.zup.pix.models.ChavePix
 import br.com.zup.pix.repositories.ChaveRepository
 import br.com.zup.pix.shared.exception.ClienteNotFoundException
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import org.slf4j.LoggerFactory
+import java.lang.IllegalStateException
 import java.util.*
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -15,7 +19,8 @@ import javax.validation.Valid
 @Validated
 @Singleton
 class RemoveChaveService(
-    @Inject val chaveRepository: ChaveRepository
+    @Inject val chaveRepository: ChaveRepository,
+    @Inject val bcbClient: BcbClient
 ) {
 
     private val logger = LoggerFactory.getLogger(this::class.java)
@@ -23,12 +28,21 @@ class RemoveChaveService(
         logger.info("Service -> Começou o processo de exclusão da chava ${request.pixId}")
 
 
-        val chavePix: Optional<ChavePix> = chaveRepository.findByIdAndClienteId(UUID.fromString(request.pixId), UUID.fromString(request.clienteId))
-        if(!chavePix.isPresent) {
+        val chavePix: Optional<ChavePix> =
+            chaveRepository.findByIdAndClienteId(UUID.fromString(request.pixId), UUID.fromString(request.clienteId))
+        if (!chavePix.isPresent) {
             logger.warn("A chave de id ${request.pixId} com cliente ${request.clienteId} não foi encontrada")
             throw ClienteNotFoundException("Não foi possivel encontrar a chave solicitada")
         }
 
+        val bcbExcluiu = bcbClient.deleta(
+            key = chavePix.get().chave,
+            request = DeletaChaveBcbRequest(key = request.pixId)
+        )
+
+        if(bcbExcluiu.status != HttpStatus.OK) {
+            throw IllegalStateException("Erro ao remover chave pix no banco central (BCB)")
+        }
         chaveRepository.delete(chavePix.get())
 
         logger.info("Service -> Terminou o processo de exclusão da chava ${request.pixId}")
